(GH-17) Add author to puppet-content-template
Prior to this commit, the `puppet-content-template` template did not
scaffold the mandatory `author` key and value into the generated
`pct-config.yml` file.

This commit ensures that it does, adding `author` as a parameter to
the template with a default value instructing the template writer to
replace it with their own author id.

Resolves #17